separate legend and layers
