+++ packages/geoview-core/public/index.html
@@ -31,7 +31,8 @@
     <h4>Basic Raw Components</h4>
     <a class="page-link" href="./raw-feature-info-1.html">Basic Feature Info (in footer panel)</a>
     <a class="page-link" href="./raw-data-table.html">Basic Data Table</a>
-    <a class="page-link" href="./raw-legend-2.html">Basic Legend</a>
+    <a class="page-link" href="./raw-legend-2.html">Basic Legend(2)</a>
+    <a class="page-link" href="./raw-layers.html">Basic Layers</a>
     <h4>Basic Loading</h4>
     <a class="page-link" href="./default-config.html">Default Configuration</a>
     <a class="page-link"
