+++ packages/geoview-core/src/core/components/layers/right-panel/WMS-style-item.tsx
