+++ packages/geoview-core/src/core/components/legend-2/legend-api.ts
@@ -1,19 +1,7 @@
 import { createElement } from 'react';
 import { LegendsLayerSet, api } from '@/app';
-import { LegendItem } from './legend-item';
-import { TypeLegendItemProps, TypeLegendProps } from './types';
-import { Legend2 } from './legend';
-
-// import { List } from '@/ui';
-
-// export interface TypeLegendProps {
-//   layerIds: string[];
-//   isRemoveable?: boolean;
-//   canSetOpacity?: boolean;
-//   expandAll?: boolean;
-//   hideAll?: boolean;
-//   canZoomTo?: boolean;
-// }
+import { TypeLegendProps } from '../layers/types';
+import { Legend } from './legend';
 
 /**
  * API to manage legend component
@@ -41,36 +29,9 @@ export class Legend2Api {
    *
    */
   createLegend = (props: TypeLegendProps) => {
-    // const { layerIds, isRemoveable, canSetOpacity, expandAll, hideAll } = props;
-    // const legendItems = layerIds.map((layerId) => {
-    //   const geoviewLayerInstance = api.maps[this.mapId].layer.geoviewLayers[layerId];
-    //   if (geoviewLayerInstance) {
-    //     return createElement(LegendItem, {
-    //       key: `layerKey-${layerId}`,
-    //       layerId,
-    //       geoviewLayerInstance,
-    //       isRemoveable,
-    //       canSetOpacity,
-    //       expandAll,
-    //       hideAll,
-    //       canZoomTo: true,
-    //     });
-    //   }
-    //   return null;
-    return createElement(Legend2, {
+    return createElement(Legend, {
       ...props,
       mapId: this.mapId,
     });
-    // return createElement('div', {}, createElement(List, { sx: { width: '100%' } }, legendItems));
-  };
-
-  /**
-   * Create an individual legend item
-   *
-   */
-  createLegendItem = (props: TypeLegendItemProps) => {
-    const { layerId } = props;
-    const geoviewLayerInstance = api.maps[this.mapId].layer.geoviewLayers[layerId];
-    return createElement(LegendItem, { layerId, geoviewLayerInstance, key: layerId });
   };
 }
