+++ packages/geoview-core/src/core/components/legend-2/legend-layer.tsx
@@ -14,8 +14,8 @@ import {
   List,
   KeyboardArrowUpIcon,
 } from '@/ui';
-import { TypeLegendLayer } from '../types';
-import { getSxClasses } from './legend-overview-styles';
+import { TypeLegendLayer } from '../layers/types';
+import { getSxClasses } from './legend-styles';
 
 interface LegendLayerProps {
   layer: TypeLegendLayer;
